+++ source/include/Ostap/Params.h
@@ -24,6 +24,7 @@ namespace Ostap
     class LegendreSum3 ;
     class LegendreSum4 ;
     class ChebyshevSum ;
+    class Bernstein    ;
     // ========================================================================
   } //                                         The end of namespace Ostap::Math
   // ==========================================================================
@@ -264,7 +265,7 @@ namespace Ostap
     // =========================================================================
     /** fill Chebyshev sum with data from the Tree 
      *  @see Ostap::Math::ChebyshevSum 
-     *  @see Ostap::Math::chebyshevSum::fill
+     *  @see Ostap::Math::ChebyshevSum::fill
      *  @param tree       (INPUT)  the input tree 
      *  @param sum        (UPDATE) the parameterization object 
      *  @param expression (INPUT)  expression to be parameterized
@@ -288,7 +289,7 @@ namespace Ostap
     // ========================================================================
     /** fill Chebyshev sum with data from the Tree 
      *  @see Ostap::Math::ChebyshevSum 
-     *  @see Ostap::Math::chebyshevSum::fill
+     *  @see Ostap::Math::ChebyshevSum::fill
      *  @param tree       (INPUT)  the input tree 
      *  @param sum        (UPDATE) the parameterization object 
      *  @param expression (INPUT)  expression to be parameterized
@@ -307,6 +308,58 @@ namespace Ostap
     static double parameterize 
     ( TTree*                     tree       , 
       Ostap::Math::ChebyshevSum& sum        , 
+      const std::string&         expression , 
+      const std::string&         selection  , 
+      const unsigned long        first      =  0 ,
+      const unsigned long        last       = std::numeric_limits<unsigned long>::max() ) ;
+    // =======================================================================
+  public:
+    // =========================================================================
+    /** fill Bernstein sum with data from the Tree 
+     *  @see Ostap::Math::Bernstein 
+     *  @see Ostap::Math::Bernstein::fill
+     *  @param tree       (INPUT)  the input tree 
+     *  @param sum        (UPDATE) the parameterization object 
+     *  @param expression (INPUT)  expression to be parameterized
+     *  @param first      (INPUT)  the first event in Tree 
+     *  @param last       (INPUT)  the last  event in Tree 
+     *  @return number of events used in parameterization 
+     *  @code
+     *  Tree*  tree = ...
+     *  Bernstein s ( 5 , -1 , 1 ) ;
+     *  DataParam::parameterize ( tree , s , "x" ) ;
+     *  @endcode
+     *  @author Vanya BELYAEV Ivan.Belyaev@itep.ru
+     *  @date   2019-07-3
+     */ 
+    static unsigned long parameterize 
+    ( TTree*                     tree       , 
+      Ostap::Math::Bernstein&    sum        , 
+      const std::string&         expression , 
+      const unsigned long        first      =  0 ,
+      const unsigned long        last       = std::numeric_limits<unsigned long>::max() ) ;
+    // ========================================================================
+    /** fill Bernstein sum with data from the Tree 
+     *  @see Ostap::Math::Bernstein 
+     *  @see Ostap::Math::Bernstein::fill
+     *  @param tree       (INPUT)  the input tree 
+     *  @param sum        (UPDATE) the parameterization object 
+     *  @param expression (INPUT)  expression to be parameterized
+     *  @param seelction  (INPUT)  selection/weight to be used 
+     *  @param first      (INPUT)  the first event in Tree 
+     *  @param last       (INPUT)  the last  event in Tree 
+     *  @return  sum of weigths  used in parameterization
+     *  @code
+     *  Tree*  tree = ...
+     *  Bernstein s ( 5 , -1 , 1 ) ;
+     *  DataParam::parameterize ( tree , s , "x"  , "y>10" ) ;
+     *  @endcode
+     *  @author Vanya BELYAEV Ivan.Belyaev@itep.ru
+     *  @date   2019-07-3
+     */ 
+    static double parameterize 
+    ( TTree*                     tree       , 
+      Ostap::Math::Bernstein&    sum        , 
       const std::string&         expression , 
       const std::string&         selection  , 
       const unsigned long        first      =  0 ,